refact: Move field methods to mixin classes
